fix circular imports for guidance links
