+++ src/DetailsView/components/issues-table-handler.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 import { IObjectWithKey } from 'office-ui-fabric-react';
 import { IGroup } from 'office-ui-fabric-react';
-import { HyperlinkDefinition } from 'views/content/content-page';
-import { RuleResult } from '../../scanner/iruleresults';
+import { HyperlinkDefinition } from 'common/types/hyperlink-definition';
+import { RuleResult } from 'scanner/iruleresults';
 
 export interface DetailsRowData extends IObjectWithKey, AxeNodeResult {
     selector: string;
