+++ src/DetailsView/components/overview-content/help-links.tsx
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as React from 'react';
-import { HyperlinkDefinition } from 'views/content/content-page';
+import { ExternalLink, ExternalLinkDeps } from 'common/components/external-link';
+import { NamedFC } from 'common/react/named-fc';
+import { HyperlinkDefinition } from 'common/types/hyperlink-definition';
 
-import { ExternalLink, ExternalLinkDeps } from '../../../common/components/external-link';
-import { NamedFC } from '../../../common/react/named-fc';
 import * as styles from './help-links.scss';
 
 export type HelpLinksDeps = ExternalLinkDeps;
