+++ src/DetailsView/components/overview-content/overview-content-container.tsx
@@ -1,17 +1,19 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { AssessmentsProvider } from 'assessments/types/assessments-provider';
 import * as React from 'react';
+
+import { AssessmentsProvider } from 'assessments/types/assessments-provider';
+import { NamedFC } from 'common/react/named-fc';
+import { HyperlinkDefinition } from 'common/types/hyperlink-definition';
+import { AssessmentStoreData } from 'common/types/store-data/assessment-result-data';
+import { FeatureFlagStoreData } from 'common/types/store-data/feature-flag-store-data';
+import { TabStoreData } from 'common/types/store-data/tab-store-data';
+import { DetailsViewActionMessageCreator } from 'DetailsView/actions/details-view-action-message-creator';
+
 import { OverviewSummaryReportModel } from 'reports/assessment-report-model';
 import { AssessmentReportSummary } from 'reports/components/assessment-report-summary';
 import { GetAssessmentSummaryModelFromProviderAndStoreData } from 'reports/get-assessment-summary-model';
-import { HyperlinkDefinition } from 'views/content/content-page';
 
-import { NamedFC } from '../../../common/react/named-fc';
-import { AssessmentStoreData } from '../../../common/types/store-data/assessment-result-data';
-import { FeatureFlagStoreData } from '../../../common/types/store-data/feature-flag-store-data';
-import { TabStoreData } from '../../../common/types/store-data/tab-store-data';
-import { DetailsViewActionMessageCreator } from '../../actions/details-view-action-message-creator';
 import { TargetChangeDialog, TargetChangeDialogDeps } from '../target-change-dialog';
 import * as styles from './overview-content-container.scss';
 import { OverviewHeading } from './overview-heading';
