+++ src/DetailsView/components/overview-content/overview-help-section.tsx
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as React from 'react';
-import { HyperlinkDefinition } from 'views/content/content-page';
 
-import { NamedFC } from '../../../common/react/named-fc';
+import { NamedFC } from 'common/react/named-fc';
+import { HyperlinkDefinition } from 'common/types/hyperlink-definition';
 import { HelpLinks, HelpLinksDeps } from './help-links';
 import * as styles from './overview-help-section.scss';
