+++ src/assessments/types/requirement.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { UniquelyIdentifiableInstances } from 'background/instance-identifier-generator';
+import { HyperlinkDefinition } from 'common/types/hyperlink-definition';
 import { ManualTestStatus } from 'common/types/manual-test-status';
 import {
     AssessmentNavState,
@@ -21,7 +22,7 @@ import { Drawer } from 'injected/visualization/drawer';
 import { DrawerProvider } from 'injected/visualization/drawer-provider';
 import { IColumn } from 'office-ui-fabric-react';
 import { DictionaryStringTo } from 'types/common-types';
-import { ContentPageComponent, HyperlinkDefinition } from 'views/content/content-page';
+import { ContentPageComponent } from 'views/content/content-page';
 import { IGetMessageGenerator } from '../assessment-default-message-generator';
 import { InstanceTableColumn } from './instance-table-column';
 import { ReportInstanceFields } from './report-instance-field';
