+++ src/common/get-guidance-tags-from-guidance-links.ts
@@ -2,8 +2,7 @@
 // Licensed under the MIT License.
 import { isArray, isObject } from 'lodash';
 
-import { GuidanceTag } from 'content/guidance-tags';
-import { GuidanceLink } from '../scanner/rule-to-links-mappings';
+import { GuidanceLink, GuidanceTag } from 'common/guidance-links';
 
 export type GetGuidanceTagsFromGuidanceLinks = (links: GuidanceLink[]) => GuidanceTag[];
 export const GetGuidanceTagsFromGuidanceLinks: GetGuidanceTagsFromGuidanceLinks = links => {
