+++ src/common/types/store-data/card-view-model.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { HighlightState } from 'common/components/cards/instance-details-footer';
-
-import { GuidanceLink } from '../../../scanner/rule-to-links-mappings';
+import { GuidanceLink } from 'common/guidance-links';
 import { InstanceResultStatus, UnifiedResult } from './unified-data-interface';
 
 export type CardRuleResultStatus = InstanceResultStatus | 'inapplicable';
