+++ src/scanner/result-decorator.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 import * as Axe from 'axe-core';
 
-import { HyperlinkDefinition } from 'views/content/content-page';
-import { DictionaryStringTo } from '../types/common-types';
+import { HyperlinkDefinition } from 'common/types/hyperlink-definition';
+import { DictionaryStringTo } from 'types/common-types';
 import { DocumentUtils } from './document-utils';
 import { AxeRule, RuleResult, ScanResults } from './iruleresults';
 import { MessageDecorator } from './message-decorator';
