+++ tsconfig.strictNullChecks.json
@@ -146,6 +146,7 @@
         "./src/common/file-url-provider.ts",
         "./src/common/get-inner-text-from-jsx-element.ts",
         "./src/common/globalization.ts",
+        "./src/common/guidance-links.ts",
         "./src/common/html-element-utils.ts",
         "./src/common/is-supported-browser.ts",
         "./src/common/itab.ts",
@@ -163,6 +164,7 @@
         "./src/common/types/deep-partial.ts",
         "./src/common/types/details-view-pivot-type.ts",
         "./src/common/types/dev-tools-open-message.ts",
+        "./src/common/types/hyperlink-definition.ts",
         "./src/common/types/idefault-constructor.ts",
         "./src/common/types/link-component-type.ts",
         "./src/common/types/manual-test-status.ts",
@@ -189,7 +191,6 @@
         "./src/common/url-parser.ts",
         "./src/common/url-validator.ts",
         "./src/common/window-utils.ts",
-        "./src/content/guidance-tags.ts",
         "./src/content/test/contrast/state-changes.tsx",
         "./src/debug-tools/action-creators/debug-tools-action-creator.ts",
         "./src/debug-tools/components/stores-tree.tsx",
